EUI-8728 Register-other-org - personal regulator
