+++ src/register-org/components/individual-registered-with-regulator-details/individual-registered-with-regulator-details.component.spec.ts
@@ -1,0 +1,40 @@
+import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { Router } from '@angular/router';
+import { RouterTestingModule } from '@angular/router/testing';
+import { EnvironmentService } from '../../../shared/services/environment.service';
+import { IndividualRegisteredWithRegulatorDetailsComponent } from './individual-registered-with-regulator-details.component';
+
+describe('IndividualRegisteredWithRegulatorDetailsComponent', () => {
+  let component: IndividualRegisteredWithRegulatorDetailsComponent;
+  let fixture: ComponentFixture<IndividualRegisteredWithRegulatorDetailsComponent>;
+
+  const mockRouter = {
+    navigate: jasmine.createSpy('navigate')
+  };
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [IndividualRegisteredWithRegulatorDetailsComponent],
+      imports: [
+        HttpClientTestingModule,
+        RouterTestingModule
+      ],
+      providers: [
+        EnvironmentService,
+        { provide: Router, useValue: mockRouter }
+      ]
+    })
+      .compileComponents();
+  });
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(IndividualRegisteredWithRegulatorDetailsComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
