+++ src/register-org/components/regulator-details/regulator-details.component.html
@@ -48,7 +48,7 @@
 
             <div *ngIf="regulators.controls[i].get('organisationRegistrationNumber')" class="govuk-!-margin-top-4">
               <label class="govuk-label govuk-!-font-weight-bold" for="'organisation-registration-number' + i">
-                Enter your organisation's registration number (optional)
+                Enter your organisation's registration number
               </label>
               <input class="govuk-input govuk-!-width-two-thirds" type="text" [id]="'organisation-registration-number' + i"
                 name="organisationRegistrationNumber" formControlName="organisationRegistrationNumber">
